Implement authorization by plugins

+++ src/Http/Controllers/BreadBuilderController.php
@@ -32,6 +32,7 @@ class BreadBuilderController extends Controller
      */
     public function index()
     {
+        $this->authorize('browse', 'breads');
         $tables = VoyagerFacade::getTables();
 
         return view('voyager::builder.index', compact('tables'));
@@ -46,6 +47,8 @@ class BreadBuilderController extends Controller
      */
     public function create($table)
     {
+        $this->authorize('create', 'bread');
+
         if (!in_array($table, VoyagerFacade::getTables())) {
             throw new \Voyager\Admin\Exceptions\TableNotFoundException('Table "'.$table.'" does not exist');
         }
@@ -85,6 +88,8 @@ class BreadBuilderController extends Controller
             return redirect()->route('voyager.bread.create', $table);
         }
 
+        $this->authorize('edit', $bread);
+
         $new = false;
 
         return view('voyager::builder.edit-add', compact('bread', 'new'));
@@ -100,6 +105,8 @@ class BreadBuilderController extends Controller
      */
     public function update(Request $request, $table)
     {
+        $this->authorize('edit', $this->breadmanager->getBreadByTable($table));
+
         $bread = $request->bread;
 
         if (!is_array($bread)) {
@@ -151,6 +158,7 @@ class BreadBuilderController extends Controller
         if (is_null($bread)) {
             return response('', 500);
         }
+        $this->authorize('delete', $bread);
         event(new DeletedEvent($bread));
 
         return response('', $this->breadmanager->deleteBread($table) ? 200 : 500);
@@ -193,6 +201,7 @@ class BreadBuilderController extends Controller
      */
     public function getBreads()
     {
+        $this->authorize('browse', 'breads');
         return response()->json([
             'breads'  => $this->breadmanager->getBreads(),
             'backups' => $this->breadmanager->getBackups(),
@@ -209,6 +218,7 @@ class BreadBuilderController extends Controller
     public function backupBread(Request $request)
     {
         $table = $request->get('table', '');
+        $this->authorize('backup', $table);
         $result = $this->breadmanager->backupBread($table);
         event(new BackedUpEvent($this->breadmanager->getBreadByTable($table)));
 
@@ -225,6 +235,7 @@ class BreadBuilderController extends Controller
     public function rollbackBread(Request $request)
     {
         $table = $request->get('table', '');
+        $this->authorize('restore', $table);
         $result = $this->breadmanager->rollbackBread($table, $request->get('path', ''));
         event(new RestoredEvent($this->breadmanager->getBreadByTable($table)));
